requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
requests==2.28.2
six==1.16.0
tabulate==0.9.0
urllib3==1.26.15
websocket-client==0.57.0

